feat(i18n): add tournament guide and various toast messages
- Added translations for tournament guide in Chinese, English, and Vietnamese.
- Introduced new toast messages for various actions in the application, including category, tournament, player, and payment operations.
- Updated the toast message handling to utilize a centralized function for consistency across services.
- Implemented a new event system for toggling the tournament setup guide visibility.

+++ src/components/tournament/TournamentCard.tsx
@@ -28,14 +28,14 @@ const formatDateRange = (startDate: Date, endDate: Date, locale: string) => {
   const start = new Date(startDate);
   const end = new Date(endDate);
   const fullFormatter = new Intl.DateTimeFormat(locale, {
-    weekday: 'short',
-    month: 'short',
+    weekday: 'long',
+    month: 'long',
     day: 'numeric',
     year: 'numeric',
   });
   const shortFormatter = new Intl.DateTimeFormat(locale, {
-    weekday: 'short',
-    month: 'short',
+    weekday: 'long',
+    month: 'long',
     day: 'numeric',
   });
   if (isSameCalendarDay(start, end)) return fullFormatter.format(start);
